Render HTML in connectedCallback instead of constructor
It is recommended to do HTML rendering inside the connectedCallback instead of
the constructor. Putting the update to `innerHTML` inside the constructor
sometimes results in a "Uncaught DOMException: Failed to construct
'CustomElement': The result must not have attributes" e.g. when the script that
creates the custom element is dynamically reloaded.

Relevant StackOverflow page: https://stackoverflow.com/a/43837330